car_order_managment/requirements.txt
django==3.2
djangorestframework==3.14.0
django-filter==22.1
django-debug-toolbar==2.2
gunicorn==20.0.4
psycopg2-binary==2.9
coverage==6.5.0